# Conflicts:
#	core/useCases/src/main/java/com/devfalah/usecases/GetUserIdUseCase.kt

+++ core/ui/src/main/java/com/devfalah/ui/screen/home/HomeScreen.kt
@@ -90,7 +90,7 @@ fun HomeContent(
     onClickLike: (PostUIState) -> Unit,
     onClickComment: (PostUIState) -> Unit,
     onClickSave: (PostUIState) -> Unit,
-    onRefresh: (Int) -> Unit,
+    onRefresh: () -> Unit,
     onClickProfile: (Int) -> Unit,
     onOpenLinkClick: (String) -> Unit,
     onDeletePost: (PostUIState) -> Unit,
